Carent - Car Rental Application (Python + MongoDB)/requirements.txt
flask
flask-pymongo
flask-login
flask-wtf
python-dotenv
passlib
pymongo
email-validator==2.0.0
bcrypt==4.0.1